Fixxing error, and modifying Toolbar
