+++ multi_tab/src/App.tsx
@@ -1,17 +1,11 @@
 import "./App.css";
-import {MapLibreMap} from "@mapcomponents/react-maplibre";
+import {MapLibreMap, TopToolbar} from "@mapcomponents/react-maplibre";
 import LayerManager from "./components/LayerManager.jsx";
 import {Button} from "@mui/material";
-import Toolbar from "@mui/material/Toolbar";
-import AppBar from "@mui/material/AppBar";
-import Box from "@mui/material/Box";
 import {useState} from "react";
 
 function App() {
     const [openSidebar, setOpenSidebar] = useState(true);
-    const ghPagesUrl = 'https://mapcomponents.github.io/react-map-components-maplibre/';
-    const logoUrl = ghPagesUrl + 'assets/WG-MapComponents-Logo_rgb.svg';
-
 
     const openTable = () => {
         const newUrl = `${window.location.href}#/table`;
@@ -19,54 +13,31 @@ function App() {
     }
     return (
         <>
-            <AppBar
-                sx={{
-                    minHeight: '62px',
-                    width: '100vw',
-                    position: 'absolute',
-                    backgroundColor: 'white',
-                    zIndex: 1300,
-                    top: 0,
-                }}
-            >
-                <Toolbar disableGutters>
-                    <Box
+            <TopToolbar buttons={
+                <>
+                    <Button
                         sx={{
-                            marginLeft: '25px',
-                            display: {xs: 'none', md: 'flex'},
-                            flexGrow: {md: '30'},
+                            color: openSidebar ? "#009EE0" : "white",
+                            backgroundColor: openSidebar ? "#fff" : '#009EE0',
+                            marginRight: {xs: '0px', sm: '10px'}
+                        }}
+                        variant={openSidebar ? 'outlined' : 'contained'}
+                        onClick={() => {
+                            setOpenSidebar(!openSidebar);
                         }}
                     >
-                        <img
-                            src={logoUrl}
-                            style={{width: '100%', maxWidth: '250px'}}
-                            alt={"MapComponentsLogo"}
-                        />
-                        <Button
-                            sx={{
-                                left: 'calc(100% - 520px)',
-                                color: '#238ee5'
-                            }}
-                            disabled={openSidebar}
-                            variant='outlined'
-                            onClick={() => {
-                                setOpenSidebar(!openSidebar);
-                            }}
-                        >
-                            Open Sidebar
-                        </Button>
+                        Sidebar
+                    </Button>
 
-                        <Button
-                            sx={{
-                                left: 'calc(100% - 510px)',
-                                color: '#238ee5'
-                            }}
-                            variant='outlined'
-                            onClick={() => openTable()}
-                        >show table</Button>
-                    </Box>
-                </Toolbar>
-            </AppBar>
+                    <Button
+                        sx={{
+                            background: '#009EE0'
+                        }}
+                        variant='contained'
+                        onClick={() => openTable()}
+                    >show table</Button>
+                </>
+            }/>
             <MapLibreMap
                 options={{
                     style: "https://wms.wheregroup.com/tileserver/style/osm-bright.json",
@@ -75,7 +46,7 @@ function App() {
                 }}
                 style={{position: "absolute", top: 0, bottom: 0, left: 0, right: 0}}
             />
-            <LayerManager setOpenSidebar={setOpenSidebar} openSidebar={openSidebar} />
+            <LayerManager setOpenSidebar={setOpenSidebar} openSidebar={openSidebar}/>
         </>
     );
 }
