+++ multi_tab/src/components/LayerManager.jsx
@@ -131,6 +131,7 @@ export default function LayerManager(props) {
 
     useEffect(() => {
         if (!mapHook.map) return;
+        console.log("send")
 
         sendMessageToServiceWorker({
             type: "visibleLayers",
